feat: add loading state in product details

+++ app/src/main/java/com/zyphir/eccomerce/angieshop/features/store/domain/mappers/ProductMapper.kt
@@ -13,7 +13,7 @@ class ProductMapper {
                     slug = item.slug,
                     name = item.name,
                     description = item.description,
-                    price = item.price,
+                    price = item.price.toInt(),
                     currency = item.currency,
                     category = item.productCategories.first().categories.name,
                     imageUrl = item.imageUrl,
